feat: refactored folder handling for multiple OS
